feat : 마이페이지 및 매장찾기 api 구현 #8

+++ src/main/java/com/togethershop/backend/domain/CouponTemplate.java
@@ -21,7 +21,7 @@ public class CouponTemplate {
     private Long templateId;
 
     @Column(name = "discount_value", nullable = false)
-    private BigDecimal discountValue;
+    private Long discountValue;
 
     @Column(name = "total_quantity")
     private Integer totalQuantity;
